+++ homelab_git_management/run.sh
@@ -38,7 +38,7 @@
 
 set -e
 
-APP_VERSION="2.1.0"
+APP_VERSION="2.2.0"
 
 export APP_VERSION